+++ migrations/migrate-drop-indexes.js
@@ -1,0 +1,1 @@
+db.deltaFile.dropIndexes(["action_search", "requeue_index"]);